Added a layout page to the settings root.
