+++ Moda.Web/src/moda.web.reactclient/src/app/settings/layout.tsx
@@ -1,0 +1,27 @@
+'use client'
+
+import { Card, Divider, Layout } from 'antd'
+import { PageTitle } from '../components/common'
+import SettingsMenu from './components/settings-menu'
+
+const { Content, Sider } = Layout
+
+const SettingsLayout = ({ children }: { children: React.ReactNode }) => {
+  return (
+    <>
+      <br />
+      <PageTitle title="Settings" />
+      <Divider />
+      <Layout>
+        <Card size="small">
+          <Sider>
+            <SettingsMenu />
+          </Sider>
+        </Card>
+        <Content style={{ paddingLeft: '24px' }}>{children}</Content>
+      </Layout>
+    </>
+  )
+}
+
+export default SettingsLayout
